# Improves perf of serialize & parse