--- a/test/perf/points.desc
+++ b/test/perf/points.desc
@@ -1,6 +1,6 @@
 
-O
-points.proto"
+Y
+test/perf/points.proto"
 Point	
 i (	
 j ("