--- a/test/test.desc
+++ b/test/test.desc
@@ -1,7 +1,6 @@
 
-þ
-
-test.prototk.tewi"ª
+ƒ
+test/test.prototk.tewi"ª
 Test
 n64 (
 name (	